pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "ghlogger"
version = "0.1.0"
description = "Tiny Gmail‑alerting Google‑Sheet logger"
authors = [{ name = "Yun jihun", email = "wlgns9835@naver.com" }]
dependencies = [
    "gspread>=5.12",
    "python-dotenv>=1.0",
]
requires-python = ">=3.9"
